test(@general): run squish test on ci

+++ scripts/ubuntu_build_setup.sh
@@ -10,13 +10,28 @@ function check_version {
   fi
 }
 
-function install_dependencies {
-  echo "Install dependencies"
+function install_build_dependencies {
+  echo "Install build dependencies"
   apt update
   apt install -yq git build-essential python3.8 python3-pip pkg-config mesa-common-dev \
      libglu1-mesa-dev wget libpcsclite-dev libpcre3-dev libssl-dev libpulse-mainloop-glib0 \
      libxkbcommon-x11-dev extra-cmake-modules cmake
+}
 
+function install_release_dependencies {
+  echo "Install release dependencies"
+  mkdir -p /usr/local/bin
+  curl -Lo/usr/local/bin/linuxdeployqt "https://github.com/probonopd/linuxdeployqt/releases/download/continuous/linuxdeployqt-continuous-x86_64.AppImage"
+  chmod a+x /usr/local/bin/linuxdeployqt
+
+  apt install -yq gstreamer1.0-plugins-good gstreamer1.0-plugins-bad \
+      gstreamer1.0-plugins-ugly gstreamer1.0-libav gstreamer1.0-tools \
+      gstreamer1.0-alsa gstreamer1.0-pulseaudio
+}
+
+function install_runtime_dependencies {
+  echo "Install runtime dependencies"
+  apt install -yq libxcomposite-dev
 }
 
 function install_qt {
@@ -27,14 +42,16 @@ function install_qt {
 }
 
 function install_golang {
-  echo "Install GoLang"
-  export GOLANG_SHA256="6e5203fbdcade4aa4331e441fd2e1db8444681a6a6c72886a37ddd11caa415d4"
-  export GOLANG_TARBALL="go1.17.12.linux-amd64.tar.gz"
-  wget -q "https://dl.google.com/go/${GOLANG_TARBALL}"
-  echo "${GOLANG_SHA256} ${GOLANG_TARBALL}" | sha256sum -c
-  tar -C /usr/local -xzf "${GOLANG_TARBALL}"
-  rm "${GOLANG_TARBALL}"
-  ln -s /usr/local/go/bin/go /usr/local/bin
+  if ! [[ -x "$(command -v go)" ]]; then
+    echo "Install GoLang"
+    export GOLANG_SHA256="006f6622718212363fa1ff004a6ab4d87bbbe772ec5631bab7cac10be346e4f1"
+    export GOLANG_TARBALL="go1.18.5.linux-arm64.tar.gz"
+    wget -q "https://dl.google.com/go/${GOLANG_TARBALL}"
+    echo "${GOLANG_SHA256} ${GOLANG_TARBALL}" | sha256sum -c
+    tar -C /usr/local -xzf "${GOLANG_TARBALL}"
+    rm "${GOLANG_TARBALL}"
+    ln -s /usr/local/go/bin/go /usr/local/bin
+  fi
 }
 
 function success_message {
@@ -44,14 +61,16 @@ SUCCESS!
 Before you attempt to build status-dektop you'll need a few environment variables set:
 
 export QTDIR=/opt/qt/5.14.2/gcc_64
-export PATH=\$PATH:\$QTDIR:\$QTDIR/bin
+export PATH=\$QTDIR:\$QTDIR/bin:\$PATH
 "
   echo $msg
 }
 
 if [ "$0" = "$BASH_SOURCE" ]; then
     check_version
-    install_dependencies
+    install_build_dependencies
+    install_release_dependencies
+    install_runtime_dependencies
     install_qt
     install_golang
     success_message
